--- Newster-news/src/main/java/com/example/Newsternews/Resources/DBUserData.java
@@ -1,46 +1,0 @@
-package com.example.Newsternews.Resources;
-
-
-import org.springframework.data.annotation.Id;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-
-@Entity
-//@Table(name = "user")
-public class DBUserData {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String firstName;
-
-    private String lastName;
-
-    @Column(unique = true)
-    private String email;
-    private String password;
-
-
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-
-}
