+++ Newster-news/src/main/java/com/example/Newsternews/Resources/SignupUserData.java
@@ -1,0 +1,25 @@
+package com.example.Newsternews.Resources;
+
+
+public class SignupUserData {
+
+    private String username;
+    private String email;
+    private String password;
+
+
+
+    public String getUsername() {
+        return username;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+
+}
